+++ src/GUI/right_panel.py
@@ -7,8 +7,10 @@ current_dir = os.path.dirname(__file__)
 parent_dir = os.path.abspath(os.path.join(current_dir, os.pardir))
 sys.path.append(parent_dir)
 
+from GUI import fetch
+
 class RightPanel(ct.CTkFrame):
-    def __init__(self, master, *args, **kwargs):
+    def __init__(self, master,terminal, *args, **kwargs):
         super().__init__(master, *args, **kwargs)
         self.configure(corner_radius=0)
         # Configure grid weights
@@ -20,7 +22,7 @@ class RightPanel(ct.CTkFrame):
         self.grid_rowconfigure(4, weight=0)
         self.grid_rowconfigure(5, weight=1)
         self.grid_rowconfigure(6, weight=0)
-
+        self.terminal = terminal
         # Label-ul pentru input
         self.input_label = ct.CTkLabel(self, text="Input Label")
         self.input_label.grid(row=0, column=0, padx=10, pady=5, sticky="n")
@@ -46,5 +48,9 @@ class RightPanel(ct.CTkFrame):
         self.expected_box.grid(row=5, column=0, padx=10, pady=(0,10), sticky="nsew")
 
         # Partea inferioara
-        self.fetch = ct.CTkButton(self,text="Fetch test cases", font=("", 16))
+        self.fetch = ct.CTkButton(self,text="Fetch test cases", font=("", 16), command=self.fetch_window)
         self.fetch.grid(row=6, column=0, padx=10, pady=(0, 5), sticky="swe")
+
+    def fetch_window(self):
+        fetch_win = fetch.FetchWindow(self.terminal, self)
+        fetch_win.mainloop()